Package extenders renamed to extender
