+++ internal/gardener/shoot/extender/provider_test.go
@@ -1,4 +1,4 @@
-package extenders
+package extender
 
 import (
 	"testing"
